requirements.txt
streamlit
python-dotenv
langchain
langchain-community
chromadb
sentence-transformers
langchain_huggingface
pypdf
langchain_chroma
